requirements.txt
fastapi==0.104.1
requests==2.31.0
python-dotenv==1.0.0
docker
pydantic==2.5.1
